requirements.txt
autopep8==1.5.7
MarkupSafe==2.0.1
numpy
openai==0.19.0
pandas==1.3.4
pandas-stubs==1.2.0.35
pycodestyle==2.8.0
python-dotenv==0.19.2
requests==2.27.0
tweepy==4.12.0
discord.py==2.1.0
PyYAML
aiohttp
async_lru==1.0.3
oauthlib==3.2.0
textblob==0.15.3
dataset
# platform
better_profanity
nltk
sqlalchemy==1.4.19
psycopg2-binary
twitter-stream.py
Django == 4.1.5
django-cors-headers == 3.13.0
djangorestframework == 3.14.0
asgiref == 3.6.0